[MOD] Moved all resources into correct folder for game project. [MOD] All dlls and resources now get copied to the build folders for all build solutions, so you can run the .exe in those folders now too.
[ADD] Observer/Subject pattern functionality added to the engine. You can now inherit from those components to make specific observer/subject functionality for your game.

[MOD] All Service Allocator based services now have defines for ease of use when making a game.

[ADD/MOD] Added a PhysicsVariables service to the engine. This contains things like the timestep and pixels per unit for Box2D. Gravity is still based on the scene itself but now has a setter function so you can change it per scene seperatly.

[MOD] In the game, instead of making a playerprefab inheriting from GameObject I know made a PlayerBehaviour 'script' inheriting from BaseComponent. This works in a similar way as if you would add a script to an object in Unity for example. This way GameObject can be made final again.

[DEL] Removed the dae namespace from certain parts of code.

+++ Minigin/SoundManager.cpp
@@ -5,11 +5,11 @@
 
 SoundManager::~SoundManager()
 {
-	for (int i{}; i < m_SoundStreams.size(); i++)
+	for (size_t i{}; i < m_SoundStreams.size(); i++)
 	{
 		delete m_SoundStreams[i];
 	}
-	for (int i{}; i < m_SoundEffects.size(); i++)
+	for (size_t i{}; i < m_SoundEffects.size(); i++)
 	{
 		delete m_SoundEffects[i];
 	}
